// src/interfaces/user.interface.ts
import { Document, Model } from "mongoose";

export interface IFullName {
    firstName: string;
    lastName: string;
}

export interface IAddress {
    street: string;
    city: string;
    country: string;
}

export interface IOrder {
    productName: string;
    price: number;
    quantity: number;
}

export interface IUser {
    userId: number;
    username: string;
    password: string;
    fullName: IFullName;
    age: number;
    email: string;
    isActive: boolean;
    hobbies: string[];
    address: IAddress;
    orders: IOrder[];
}

export interface UserModel extends Model<IUser> {
    isUserExists(userId: number): Promise<IUser | null>;
}